+++ src/lib/components/Editor/Editor.tsx
@@ -5,169 +5,15 @@ import "./index.css";
 import React, { useRef } from "react";
 import { schema } from "prosemirror-schema-basic";
 import { keymap } from "prosemirror-keymap";
-import { baseKeymap, Command, toggleMark } from "prosemirror-commands";
-import { MarkType, Schema, NodeType } from "prosemirror-model";
-import { history, redo, undo } from "prosemirror-history";
+import { baseKeymap } from "prosemirror-commands";
 import { useProseMirror, ProseMirror } from "use-prosemirror";
-import { EditorState, Plugin } from "prosemirror-state";
+import { Plugin } from "prosemirror-state";
 
-import {
-  blockTypeItem,
-  Dropdown,
-  menuBar,
-  DropdownSubmenu,
-  undoItem,
-  redoItem,
-  icons,
-  MenuItem,
-  MenuElement,
-  MenuItemSpec
-} from "prosemirror-menu";
+import { menuBar } from "prosemirror-menu";
 
-import {
-  SubscribHandler,
-  SubscriberCallback,
-  Subscriber,
-  PluginFactory,
-  BasePluginState
-} from "../../types";
+import { BasePluginState } from "../../types";
 
-import cut from "./cut";
-
-import Popup from "../Popup";
-
-function cmdItem(cmd: Command, options: Partial<MenuItemSpec>) {
-  const passedOptions: MenuItemSpec = {
-    label: options.title as string | undefined,
-    run: cmd
-  };
-  for (let prop in options)
-    (passedOptions as any)[prop] = (options as any)[prop];
-  if (!options.enable && !options.select)
-    passedOptions[options.enable ? "enable" : "select"] = state => cmd(state);
-
-  return new MenuItem(passedOptions);
-}
-
-function markActive(state: EditorState, type: MarkType) {
-  const { from, $from, to, empty } = state.selection;
-  if (empty) return !!type.isInSet(state.storedMarks || $from.marks());
-  else return state.doc.rangeHasMark(from, to, type);
-}
-
-function markItem(markType: MarkType, options: Partial<MenuItemSpec>) {
-  const passedOptions: Partial<MenuItemSpec> = {
-    active(state) {
-      return markActive(state, markType);
-    }
-  };
-  for (const prop in options)
-    (passedOptions as any)[prop] = (options as any)[prop];
-  return cmdItem(toggleMark(markType), passedOptions);
-}
-
-type MenuItemResult = {
-  /// A menu item to toggle the [strong mark](#schema-basic.StrongMark).
-  toggleStrong?: MenuItem;
-
-  /// A menu item to toggle the [emphasis mark](#schema-basic.EmMark).
-  toggleEm?: MenuItem;
-
-  /// A menu item to set the current textblock to be a normal
-  /// [paragraph](#schema-basic.Paragraph).
-  makeParagraph?: MenuItem;
-
-  /// Menu items to set the current textblock to be a
-  /// [heading](#schema-basic.Heading) of level _N_.
-  makeHead1?: MenuItem;
-  makeHead2?: MenuItem;
-  makeHead3?: MenuItem;
-  makeHead4?: MenuItem;
-  makeHead5?: MenuItem;
-  makeHead6?: MenuItem;
-
-  /// A menu item to set the current textblock to be a
-  /// [code block](#schema-basic.CodeBlock).
-  makeCodeBlock?: MenuItem;
-
-  /// Inline-markup related menu items.
-  inlineMenu: MenuElement[][];
-
-  /// A dropdown containing the items for making the current
-  /// textblock a paragraph, code block, or heading.
-  typeMenu: Dropdown;
-
-  /// An array of arrays of menu elements for use as the full menu
-  /// for, for example the [menu
-  /// bar](https://github.com/prosemirror/prosemirror-menu#user-content-menubar).
-  fullMenu: MenuElement[][];
-};
-
-/// Given a schema, look for default mark and node types in it and
-/// return an object with relevant menu items relating to those marks.
-export function buildMenuItems(schema: Schema): MenuItemResult {
-  const r: MenuItemResult = {} as any;
-  let mark: MarkType | undefined;
-  if ((mark = schema.marks.strong))
-    r.toggleStrong = markItem(mark, {
-      title: "Toggle strong style",
-      icon: icons.strong
-    });
-  if ((mark = schema.marks.em))
-    r.toggleEm = markItem(mark, { title: "Toggle emphasis", icon: icons.em });
-
-  let node: NodeType | undefined;
-
-  if ((node = schema.nodes.paragraph))
-    r.makeParagraph = blockTypeItem(node, {
-      title: "Change to paragraph",
-      label: "Plain"
-    });
-
-  if ((node = schema.nodes.heading))
-    for (let i = 1; i <= 10; i++)
-      (r as any)["makeHead" + i] = blockTypeItem(node, {
-        title: "Change to heading " + i,
-        label: "Level " + i,
-        attrs: { level: i }
-      });
-
-  r.typeMenu = new Dropdown(
-    cut([
-      r.makeParagraph,
-      r.makeCodeBlock,
-      r.makeHead1 &&
-        new DropdownSubmenu(
-          cut([
-            r.makeHead1,
-            r.makeHead2,
-            r.makeHead3,
-            r.makeHead4,
-            r.makeHead5,
-            r.makeHead6
-          ]),
-          { label: "Heading" }
-        )
-    ]),
-    { label: "Type..." }
-  );
-
-  r.inlineMenu = [cut([r.toggleStrong, r.toggleEm])];
-
-  r.fullMenu = r.inlineMenu.concat([[r.typeMenu]]);
-
-  return r;
-}
-
-const mySchema = new Schema({
-  nodes: {
-    doc: schema.nodes.doc,
-    paragraph: schema.nodes.paragraph,
-    heading: schema.nodes.heading,
-    text: schema.nodes.text
-  },
-  marks: schema.spec.marks
-});
+import buildMenu from "./buildMenu";
 
 interface EditorProps {
   plugins: Plugin<BasePluginState, any>[];
@@ -180,7 +26,7 @@ const Editor = ({ plugins, className = null }: EditorProps) => {
     plugins: [
       menuBar({
         floating: true,
-        content: buildMenuItems(schema).fullMenu
+        content: buildMenu(schema).fullMenu
       }),
       ...plugins,
